+++ docs/examples/02_pet_sinogram_projections/02_run_pet_tof_sinogram_projector.py
@@ -44,7 +44,7 @@ scanner = parallelproj.pet_scanners.RegularPolygonPETScannerGeometry(
     num_lor_endpoints_per_side=15,
     lor_spacing=2.3,
     ring_positions=xp.linspace(-4, 4, num_rings, device=dev),
-    symmetry_axis=1,
+    symmetry_axis=2,
 )
 
 # %%
@@ -69,7 +69,7 @@ lor_desc = parallelproj.pet_lors.RegularPolygonPETLORDescriptor(
 # define a first projector using an image with 40x8x40 voxels of size 2x2x2 mm
 # where the image center is at world coordinate (0, 0, 0)
 proj = parallelproj.projectors.RegularPolygonPETProjector(
-    lor_desc, img_shape=(40, 7, 40), voxel_size=(2.0, 2.0, 2.0)
+    lor_desc, img_shape=(40, 40, 7), voxel_size=(2.0, 2.0, 2.0)
 )
 
 
@@ -82,6 +82,7 @@ proj = parallelproj.projectors.RegularPolygonPETProjector(
 
 fig = plt.figure(figsize=(8, 8))
 ax1 = fig.add_subplot(111, projection="3d")
+ax1.view_init(elev=-30, azim=160, roll=180, vertical_axis="y")
 proj.show_geometry(ax1)
 fig.tight_layout()
 fig.show()